refactor(shielded): use pubkey size constant + move IShieldedAddressInfo to shielded types
Review responses:

- assertValidCompressedPubkey: the explicit shape check is NOT redundant
  with PublicKey.fromBuffer — bitcore also accepts 65-byte uncompressed
  keys (04/06/07 prefix), verified empirically, which would silently
  corrupt the fixed 33-byte slots of the 71-byte address layout; and
  bitcore's own shape error is an opaque 'Invalid X'. Keep the check,
  switch the literal 33 to COMPRESSED_PUBKEY_SIZE_BYTES, and document the
  rationale in the doc comment.
- Move IShieldedAddressInfo from utils/shieldedAddress.ts to
  src/shielded/types.ts, the established home for shielded-domain types
  since PR 2 (#1085). No stack ripple: pr-5/pr-6 use deriveShieldedAddress's
  return type structurally and never import the interface by name.

lint, build, and the full unit suite (79 suites / 1022 tests) pass.
